feat(Next > Album): Carousel component in both All and album
