+++ app/pages/[gallery]/[album].jsx
@@ -1,12 +1,11 @@
 import Head from 'next/head'
-import ImageGallery from 'react-image-gallery'
-import 'react-image-gallery/styles/css/image-gallery.css'
 import styled from 'styled-components'
 
 import { get as getAlbum } from '../../src/lib/album'
 import { get as getAlbums } from '../../src/lib/albums'
 import { get as getGalleries } from '../../src/lib/galleries'
 
+import SplitViewer from '../../src/components/SpiltViewer'
 import ThumbImg from '../../src/components/ThumbImg'
 import useSearch from '../../src/hooks/useSearch'
 
@@ -46,34 +45,6 @@ const Wrapper = styled.ul`
   padding-left: 2px;
 `
 
-const toCarousel = (item) => {
-  const imageGallery = {
-    original: item.photoPath || item.thumbPath,
-    thumbnail: item.thumbPath,
-    description: item.description,
-    filename: item.filename,
-    videoPath: item.videoPath,
-  }
-
-  return imageGallery
-}
-
-const Split = styled.section`
-    display: grid;
-    grid-template-columns: 60% 40%;
-    grid-template-areas: 'left right';
-  `
-
-const Left = styled.section`
-    grid-area: left;
-    height: 80vh;
-  `
-
-const Right = styled.section`
-    grid-area: right;
-    height: 80vh;
-  `
-
 const AlbumPage = ({ items = [] }) => {
   const {
     filtered,
@@ -87,19 +58,7 @@ const AlbumPage = ({ items = [] }) => {
         <link rel="icon" href="/favicon.ico" />
       </Head>
       {searchBox}
-      <Split>
-        <Left key="splitLeft">
-          <ImageGallery
-            items={filtered.filter((item) => item.thumbPath).map(toCarousel)}
-            showPlayButton={false}
-            showThumbnails={false}
-            slideDuration={550}
-          />
-        </Left>
-        <Right key="splitRight">
-          Right
-        </Right>
-      </Split>
+      <SplitViewer items={filtered} />
       <Wrapper>
         {filtered.map((item) => (
           <ThumbImg src={item.thumbPath} caption={item.caption} key={item.filename} id={`select${item.id}`} />
